feat: move the function about mouse to BuiltinApplet
